# Improve Mutex performance